+++ src/hooks/index.js
@@ -1,1 +1,2 @@
 export * from './useAuthorityLinking';
+export * from './useAuthoritySourceFiles';